refactor(channel): extract domain adapters from assembly root
